car_damage/requirement.txt
django==2.0.2
djangorestframework
django-debug-toolbar
djongo==1.2.29
pymongo>=3.2.0
markdown
django-filter
setuptools
Pillow
jsonfield>=2.0.2
django-rest-swagger==2.1.2
celery
torch
scipy
numpy
scikit-image 
torchvision
SimpleITK
sklearn
scikit-learn
progressbar2
visdom
../avengers_django_models/avengers_django_models_package/dist/avengers_django_models_app-0.4.tar.gz
drfapikey

